EHN add tensorflow sequence
