+++ imblearn/keras/__init__.py
@@ -1,8 +1,8 @@
 """The :mod:`imblearn.keras` provides utilities to deal with imbalanced dataset
 in keras."""
 
-from .generator import BalancedBatchGenerator
-from .generator import balanced_batch_generator
+from ._generator import BalancedBatchGenerator
+from ..tensorflow._generator import balanced_batch_generator
 
 __all__ = ['BalancedBatchGenerator',
            'balanced_batch_generator']
